refactor: merge dev and restructure remaining outlier tests
