+++ test/e2e/strategies/aave/CloseToCollateral.test.ts
@@ -8,10 +8,10 @@ import { BigNumber } from 'bignumber.js'
 import { expect } from 'chai'
 import { loadFixture } from 'ethereum-waffle'
 
-import { getSystemWithAavePositions, SystemWithAAVEPositions } from '../../fixtures'
-import { getSystemWithAaveV3Positions } from '../../fixtures/system/getSystemWithAaveV3Positions'
-import { SystemWithAAVEV3Positions } from '../../fixtures/types/systemWithAAVEPositions'
-import { expectToBeEqual } from '../../utils'
+import { getSystemWithAavePositions, SystemWithAAVEPositions } from '../../../fixtures'
+import { getSystemWithAaveV3Positions } from '../../../fixtures/system/getSystemWithAaveV3Positions'
+import { SystemWithAAVEV3Positions } from '../../../fixtures/types/systemWithAAVEPositions'
+import { expectToBeEqual } from '../../../utils'
 
 const EXPECT_DEBT_BEING_PAID_BACK = 'Expect debt being paid back'
 const EXPECT_FEE_BEING_COLLECTED = 'Expect fee being collected'
