+++ test/e2e/strategies/aave/ReopenStEth.test.ts
@@ -19,12 +19,12 @@ import { expect } from 'chai'
 import { loadFixture } from 'ethereum-waffle'
 import { Contract, Signer } from 'ethers'
 
-import { zero } from '../../../scripts/common'
-import { mainnetAddresses } from '../../addresses'
-import { testBlockNumber } from '../../config'
-import { deploySystem } from '../../deploySystem'
-import { initialiseConfig } from '../../fixtures'
-import { expectToBe, expectToBeEqual } from '../../utils'
+import { zero } from '../../../../scripts/common'
+import { mainnetAddresses } from '../../../addresses'
+import { testBlockNumber } from '../../../config'
+import { deploySystem } from '../../../deploySystem'
+import { initialiseConfig } from '../../../fixtures'
+import { expectToBe, expectToBeEqual } from '../../../utils'
 
 describe(`Strategy | AAVE | Reopen Position`, async () => {
   const depositAmountInWei = amountToWei(new BigNumber(1))
