+++ test/e2e/strategies/maker/Close.test.ts
@@ -21,10 +21,10 @@ import { loadFixture } from 'ethereum-waffle'
 import { Contract, Signer } from 'ethers'
 import { ethers } from 'hardhat'
 
-import { testBlockNumber } from '../../config'
-import { DeployedSystemInfo } from '../../deploySystem'
-import { initialiseConfig } from '../../fixtures/setup'
-import { expectToBeEqual } from '../../utils'
+import { testBlockNumber } from '../../../config'
+import { DeployedSystemInfo } from '../../../deploySystem'
+import { initialiseConfig } from '../../../fixtures/setup'
+import { expectToBeEqual } from '../../../utils'
 
 const createAction = ActionFactory.create
