+++ test/e2e/strategies/maker/IncreaseMultiple.test.ts
@@ -24,10 +24,10 @@ import { loadFixture } from 'ethereum-waffle'
 import { Contract, Signer } from 'ethers'
 import { ethers } from 'hardhat'
 
-import { testBlockNumber } from '../../config'
-import { DeployedSystemInfo } from '../../deploySystem'
-import { initialiseConfig } from '../../fixtures/setup'
-import { expectToBe, expectToBeEqual } from '../../utils'
+import { testBlockNumber } from '../../../config'
+import { DeployedSystemInfo } from '../../../deploySystem'
+import { initialiseConfig } from '../../../fixtures/setup'
+import { expectToBe, expectToBeEqual } from '../../../utils'
 
 const LENDER_FEE = new BigNumber(0)
